manage delete movie in card only if admin through list page
