+++ app/components/card.tsx
@@ -66,12 +66,12 @@ export default function Card({
                 className="object-cover w-full h-full  transform transition-transform duration-700 ease-in-out group-hover:scale-110 group-hover:brightness-50"
                 title={title}
               />
+              <div className="absolute bottom-0 left-0 w-full p-5 bg-gradient-to-t via-transparent to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300">
+                <div className="text-md font-semibold uppercase">{title}</div>
+                <p className="text-sm text-gray-300">{release_date}</p>
+                {/* <p className="absolute text-sm text-gray-200 mt-2">{description}</p> */}
+              </div>
             </Link>
-            <div className="absolute bottom-0 left-0 w-full p-5 bg-gradient-to-t via-transparent to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300">
-              <div className="text-md font-semibold uppercase">{title}</div>
-              <p className="text-sm text-gray-300">{release_date}</p>
-              {/* <p className="absolute text-sm text-gray-200 mt-2">{description}</p> */}
-            </div>
             {userIsAdmin && (
               <button
                 onClick={handleDeleteClick}
